1.1.0

+++ EliteDangerousINARA/src/EliteDangerousINARA/Commands/AddCommanderShip.cs
@@ -19,9 +19,9 @@ namespace NSW.EliteDangerous.INARA.Commands
         public string ShipType { get; internal set; }
 
         [JsonProperty("shipGameID")]
-        public int ShipId { get; internal set; }
+        public long ShipId { get; internal set; }
 
-        public AddCommanderShip(string shipType, int shipId)
+        public AddCommanderShip(string shipType, long shipId)
         {
             if(string.IsNullOrWhiteSpace(shipType)) throw new ArgumentNullException(nameof(shipType));
             if(shipId<0) throw new ArgumentOutOfRangeException(nameof(shipId));
